Dynamic Ordering
Fixed the sorting and rendering of things.
