+++ UUG01/Assets/Prefabs/FarmHouse.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 6094949954313566865}
   - component: {fileID: -5261362554727312041}
   - component: {fileID: 6198296127849522888}
+  - component: {fileID: 6634573903049114922}
   m_Layer: 9
   m_Name: FarmHouse
   m_TagString: Untagged
@@ -158,3 +159,15 @@ PolygonCollider2D:
       - {x: 1.5470657, y: 0.7202401}
       - {x: 1.1330171, y: 0.72646165}
   m_UseDelaunayMesh: 0
+--- !u!114 &6634573903049114922
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 7707809817765145238}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2cc4e634c6577c0408b06694e48b58be, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
