+++ UUG01/Assets/Prefabs/House.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 6481756279212730488}
   - component: {fileID: 6481756279212730487}
   - component: {fileID: 977922022683107016}
+  - component: {fileID: -8447197289266516094}
   m_Layer: 9
   m_Name: House
   m_TagString: Untagged
@@ -142,3 +143,15 @@ PolygonCollider2D:
       - {x: 1.9921875, y: -0.2359209}
       - {x: 2.0234375, y: 3.8758411}
   m_UseDelaunayMesh: 0
+--- !u!114 &-8447197289266516094
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6481756279212730486}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2cc4e634c6577c0408b06694e48b58be, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
