+++ UUG01/Assets/Prefabs/MarketHouse2.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 9012897703146508060}
   - component: {fileID: -256854388577785122}
   - component: {fileID: 8004063995164500082}
+  - component: {fileID: -6748618101856772859}
   m_Layer: 9
   m_Name: MarketHouse2
   m_TagString: Untagged
@@ -152,3 +153,15 @@ PolygonCollider2D:
       - {x: 2.5716414, y: 3.2650735}
       - {x: 2.541259, y: 3.5870204}
   m_UseDelaunayMesh: 0
+--- !u!114 &-6748618101856772859
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 4884216305047946759}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2cc4e634c6577c0408b06694e48b58be, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
