--- UUG01/Assets/Prefabs/MarketHouse4.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 11222b63f9dc70f4ea5e8f88a87697da
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
