--- UUG01/Assets/Prefabs/MarketHouse5.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d24b98bf19536ea4880493a708fde421
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
