+++ UUG01/Assets/Prefabs/Well.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 6778482603490928105}
   - component: {fileID: 1384776934637804636}
   - component: {fileID: -6319703043583260314}
+  - component: {fileID: 4890822826737646853}
   m_Layer: 0
   m_Name: Well
   m_TagString: Untagged
@@ -152,3 +153,15 @@ PolygonCollider2D:
       - {x: 1.1600356, y: 0.027164713}
       - {x: 1.3755267, y: 0.28165513}
   m_UseDelaunayMesh: 0
+--- !u!114 &4890822826737646853
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 618484175776527625}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2cc4e634c6577c0408b06694e48b58be, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier:
