--- UUG01/Assets/Scripts/Crop.cs.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< Updated upstream
-guid: fb9e16723a2071d4d8850758d2ce4c3f
-=======
-guid: 45dc063f921fc524ba25359fccb3fd98
->>>>>>> Stashed changes
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
