+++ UUG01/Assets/Scripts/Crops.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: af91e898f0d08da42a10935f9463e624
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
