+++ UUG01/Assets/Scripts/Crops/HarvestSystem.cs.meta
@@ -1,0 +1,15 @@
+fileFormatVersion: 2
+<<<<<<< Updated upstream
+guid: 3c115a1ea50ee744db2c0634ec142db6
+=======
+guid: 8e2b2c21046b9b243b7bf31802d26ae6
+>>>>>>> Stashed changes
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
