--- UUG01/Assets/Scripts/DayAndNightCycle.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 0ac9dd6ff47f13f47be4d7af2ddc3b5e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
