+++ UUG01/Assets/Scripts/DynamicSorting.cs
@@ -1,0 +1,19 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEditor.U2D;
+using UnityEngine;
+
+public class NewBehaviourScript : MonoBehaviour
+{
+    private SpriteRenderer spriteRenderer;
+
+    private void Start()
+    {
+        spriteRenderer = GetComponent<SpriteRenderer>();
+    }
+
+    private void Update()
+    {
+        spriteRenderer.sortingOrder = Mathf.RoundToInt(transform.position.y * -100);
+    }
+}
