+++ UUG01/Assets/Scripts/NightAndDay.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3916802a8638e9246a3378f2cbfe5eb6
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
