+++ UUG01/Assets/Scripts/NightAndDay/WorldLight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac8a5fd87a994c74db042114269336d9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
