+++ UUG01/Assets/Scripts/NightAndDay/WorldTimeConstants.cs
@@ -1,0 +1,9 @@
+using System.Collections;
+using System.Collections.Generic;
+
+
+public static class WorldTimeConstants
+{
+
+    public const int MinutesInDay = 1440;
+}
