--- UUG01/Assets/Scripts/PlantingSystem.cs.meta
@@ -1,15 +1,0 @@
-fileFormatVersion: 2
-<<<<<<< Updated upstream
-guid: 45cdd3de31c70e54eb905e5aab741252
-=======
-guid: 5792bdd4f889b194588a1f0091fe84d4
->>>>>>> Stashed changes
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
