--- UUG01/Assets/Scripts/WorldTime.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 6b6c9fdcfd0e5e849b07e8da717cb509
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
