--- UUG01/Assets/Scripts/WorldTimeConstants.cs
@@ -1,9 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-
-
-public static class WorldTimeConstants
-{
-
-    public const int MinutesInDay = 1440;
-}
