--- UUG01/Assets/Scripts/WorldTimeConstants.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4407b2d837ab70148a3241b0e9251abf
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
